Fixing build for x86_64-w64-mingw32

+++ src/gtk/widgets/treepath.rs
@@ -53,7 +53,7 @@ impl TreePath {
     }
 
     pub fn new_from_indicesv(indices: &mut [i32]) -> Option<TreePath> {
-        let tmp = unsafe { ffi::gtk_tree_path_new_from_indicesv(indices.as_mut_ptr(), indices.len() as u64) };
+        let tmp = unsafe { ffi::gtk_tree_path_new_from_indicesv(indices.as_mut_ptr(), indices.len() as u32) };
 
         if tmp.is_null() {
             None
